Rewrote using c++ style, fixed issue with negative time in config
